requirements.txt
# Python dependencies for README generation

# GitHub API interaction
requests>=2.31.0
PyGithub>=2.1.1

# Date/time handling
python-dateutil>=2.8.2

